pyproject.toml
[tool.poetry]
name = "el Bandolero"
version = "2.0.0"
description = ""
authors = ["Duarte Leitão"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
discord = "^1.7.3"
Flask = "^2.1.2"
Pillow = "^10.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"